Refactor: module extraction
